requirements.txt
# Requires Python 3.11 (tested with CPython 3.11.x on Windows)
Flask>=2.3.3,<3.1.0
python-dotenv>=1.0.0,<2.0.0
supabase>=2.4.0,<3.0.0
ollama>=0.1.7,<0.2.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0,<11.0.0
tensorflow>=2.15.0,<2.17.0
scikit-learn>=1.3.0,<2.0.0
pandas>=2.1.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.13.0,<0.14.0
streamlit>=1.28.0,<2.0.0
librosa>=0.10.1,<0.11.0
ipykernel>=6.25.0,<7.0.0
opencv-python>=4.8.0,<5.0.0
networkx>=3.1.0,<4.0.0
altair>=5.0.0,<6.0.0
scipy>=1.11.0,<2.0.0
xgboost>=1.7.6,<2.0.0
joblib>=1.3.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
markdown>=3.5.0,<4.0.0
PyPDF2>=3.0.0,<4.0.0
unstructured>=0.10.30,<0.11.0
